+++ scripts/build-benchmarks/build-benchmarks.spec.js
@@ -1,5 +1,0 @@
-import { buildBenchmarks } from './build-benchmarks'
-
-test('happy', async () => {
-  await buildBenchmarks()
-})